refactor: remove RequestsAdminPage.module.css
