--- frontend/src/components/pages/RequestsAdminPage/RequestsAdminPage.module.css
@@ -1,10 +1,0 @@
-.apply-filters-container {
-  display: grid;
-  grid-template-columns: repeat(2, auto);
-  grid-template-rows: repeat(2, auto);
-
-  justify-content: start;
-  align-items: center;
-
-  gap: 5px 20px;
-}
